Create PROFILE.md
I have added the PROFILE.md file to the directory. This file can be used to provide information about the user or project. It is a good practice to include a profile file in any project to provide context and information for others who may be working on the project.